pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.2.3",
    "openai>=1.63.2",
    "pillow>=11.1.0",
    "scikit-learn>=1.6.1",
]
